+++ docs/source/_static/CHANGELOG.rst
@@ -11,6 +11,25 @@ See the `releases for ts-backend-check <https://github.com/activist-org/ts-backe
 
 Emojis for the following are chosen based on `gitmoji <https://gitmoji.dev/>`_.
 
+ts-backend-check 1.2.1
+----------------------
+
+♻️ Code Refactoring
+^^^^^^^^^^^^^^^^^^^
+
+- The comment to ignore fields within TypeScript interfaces has been updated to ``// ts-backend-check: ignore field {fieldName}``.
+- The output messages were updated to improve clarity with regards to file names being compared as well as word choice.
+
+✅ Tests
+^^^^^^^^
+
+- Tests were updated to account for the code refactoring.
+
+📝 Documentation
+^^^^^^^^^^^^^^^^
+
+- The readme and documentation files were updated to show the functionality of the package (`#20 <https://github.com/activist-org/ts-backend-check/issues/20>`_).
+
 ts-backend-check 1.2.0
 ----------------------